chore: refine pre-estimate fee UI  (#531)
* chore: handle comments

* chore: fix showText should be optional in loadingSmallView

* chore: use new modal

* chore: fix comments

* chore: fix comments

* chore: fix comments

* chore: fix use estimate fee should use fields.feeToken
