+++ packages/wallet-ui/src/components/ui/atom/LoadingSmall/LoadingSmall.view.tsx
@@ -2,15 +2,17 @@ import { HTMLAttributes } from 'react';
 import { LoadingSpinner, LoadingText, Wrapper } from './LoadingSmall.style';
 import { useMultiLanguage } from 'services';
 
-interface Props extends HTMLAttributes<HTMLDivElement> {}
+interface Props extends HTMLAttributes<HTMLDivElement> {
+  showText?: boolean;
+}
 
-export const LoadingSmallView = ({ ...otherProps }: Props) => {
+export const LoadingSmallView = ({ showText = true, ...otherProps }: Props) => {
   const { translate } = useMultiLanguage();
 
   return (
     <Wrapper {...otherProps}>
       <LoadingSpinner icon="spinner" pulse />
-      <LoadingText>{translate('loading')}</LoadingText>
+      {showText && <LoadingText>{translate('loading')}</LoadingText>}
     </Wrapper>
   );
 };
