+++ packages/wallet-ui/src/components/ui/molecule/AmountInput/AmountInput.stories.tsx
@@ -19,11 +19,11 @@ const asset = {
 };
 
 export const Default = () => (
-  <AmountInputView label="Amount" asset={asset} isFetchingFee={false} />
+  <AmountInputView label="Amount" asset={asset} isEstimatingGas={false} />
 );
 
 export const ErrorView = () => (
-  <AmountInputView error label="Amount" asset={asset} isFetchingFee={false} />
+  <AmountInputView error label="Amount" asset={asset} isEstimatingGas={false} />
 );
 
 export const Disabled = () => (
@@ -31,7 +31,7 @@ export const Disabled = () => (
     disabled
     label="Amount"
     asset={asset}
-    isFetchingFee={false}
+    isEstimatingGas={false}
   />
 );
 
@@ -41,6 +41,6 @@ export const WithHelperText = () => (
     helperText="Helper text"
     label="Amount"
     asset={asset}
-    isFetchingFee={false}
+    isEstimatingGas={false}
   />
 );
