+++ packages/wallet-ui/src/components/ui/molecule/AmountInput/AmountInput.view.tsx
@@ -11,7 +11,7 @@ import {
   RowWrapper,
   USDDiv,
   Wrapper,
-  Spinner,
+  LoadingWrapper,
 } from './AmountInput.style';
 import { Erc20TokenBalance } from 'types';
 import { ethers } from 'ethers';
@@ -25,9 +25,9 @@ interface Props extends InputHTMLAttributes<HTMLInputElement> {
   decimalsMax?: number;
   asset: Erc20TokenBalance;
   onChangeCustom?: (value: string) => void;
-  isFetchingFee: boolean;
-  shouldApplyMax?: boolean;
-  setShouldApplyMax?: (value: boolean) => void;
+  isEstimatingGas: boolean;
+  isMaxAmountPending?: boolean;
+  setIsMaxAmountPending?: (value: boolean) => void;
 }
 
 export const AmountInputView = ({
@@ -39,9 +39,9 @@ export const AmountInputView = ({
   decimalsMax = 18,
   asset,
   onChangeCustom,
-  isFetchingFee,
-  shouldApplyMax = false,
-  setShouldApplyMax,
+  isEstimatingGas,
+  isMaxAmountPending = false,
+  setIsMaxAmountPending,
   ...otherProps
 }: Props) => {
   const [focused, setFocused] = useState(false);
@@ -93,13 +93,16 @@ export const AmountInputView = ({
 
   useEffect(
     () => {
-      if (shouldApplyMax && !isFetchingFee) {
+      // If isMaxAmountPending is true and the fees are not being fetched,
+      // apply the maximum amount and reset isMaxAmountPending to false.
+      if (isMaxAmountPending && !isEstimatingGas) {
         handleMaxClick();
-        typeof setShouldApplyMax === 'function' && setShouldApplyMax(false);
+        typeof setIsMaxAmountPending === 'function' &&
+          setIsMaxAmountPending(false);
       }
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
-    [shouldApplyMax, isFetchingFee],
+    [isMaxAmountPending, isEstimatingGas],
   );
 
   const handleContainerClick = () => {
@@ -109,8 +112,11 @@ export const AmountInputView = ({
   };
 
   const handleMaxClick = () => {
-    if (isFetchingFee) {
-      typeof setShouldApplyMax === 'function' && setShouldApplyMax(true);
+    // If the fees are being fetched, set isMaxAmountPending to true
+    // to apply the maximum amount once the fees are fetched.
+    if (isEstimatingGas) {
+      typeof setIsMaxAmountPending === 'function' &&
+        setIsMaxAmountPending(true);
       return;
     }
     let amountBN = ethers.BigNumber.from(asset.amount);
@@ -159,7 +165,7 @@ export const AmountInputView = ({
             onChange={(event) => triggerOnChange(event.target.value)}
             {...otherProps}
           />
-          {isFetchingFee && shouldApplyMax && <Spinner />}
+          {isEstimatingGas && isMaxAmountPending && <LoadingWrapper />}
           {!usdMode && (
             <>
               {asset.symbol}
