+++ packages/wallet-ui/src/components/ui/organism/Header/SendInputModal/index.ts
@@ -1,0 +1,1 @@
+export { SendInputModalView as SendInputModal } from './SendInputModal.view';
