+++ packages/wallet-ui/src/components/ui/organism/Header/SendSummaryModal/SendSummaryModal.style.ts
@@ -3,29 +3,6 @@ import { Button } from 'components/ui/atom/Button';
 import { LoadingSmall } from 'components/ui/atom/LoadingSmall';
 import { Alert } from 'components/ui/atom/Alert';
 
-export const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  background-color: ${(props) => props.theme.palette.grey.white};
-  width: ${(props) => props.theme.modal.base};
-  padding: ${(props) => props.theme.spacing.base};
-  border-radius: 8px 8px 0px 0px;
-`;
-
-export const Header = styled.div`
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-`;
-
-export const Title = styled.div`
-  font-size: ${(props) => props.theme.typography.h3.fontSize};
-  font-weight: ${(props) => props.theme.typography.h3.fontWeight};
-  font-family: ${(props) => props.theme.typography.h3.fontFamily};
-  width: 100%;
-  text-align: center;
-`;
-
 export const ToDiv = styled.div`
   font-size: ${(props) => props.theme.typography.p2.fontSize};
   font-weight: 900;
